+++ spectacle/models/__init__.py
@@ -1,0 +1,3 @@
+from .converters import *
+from .custom import *
+from .profiles import *
