+++ spectacle/utils/spell.py
